move logger files to a folder
