+++ src/config/Config.h
@@ -45,7 +45,7 @@
 #include <boost/property_tree/ptree.hpp>
 #include <boost/property_tree/xml_parser.hpp>
 
-#include "../logger.h"
+#include "../logger/logger.h"
 
 using namespace std;
 using boost::property_tree::ptree;
